+++ packages/analytics_hub_appsflyer/example/main.dart
@@ -44,11 +44,8 @@ Future<void> main() async {
     ],
   );
 
-  await hub.initialize();
-
   await hub.sendEvent(
     const ExampleEvent(exampleProperty: 'example_property'),
   );
 
-  await hub.dispose();
 }